+++ packages/core/src/Footer/index.js
@@ -33,6 +33,6 @@ export default Footer;
 // Expose components
 export { default as AboutOrganization } from "../AboutOrganization";
 export { default as Copyright } from "../Copyright";
-export { default as FooterInitiativeLogo } from "./Initiative";
+export { default as InitiativeLogo } from "../InitiativeLogo";
 export { default as LegalLinks } from "../LegalLinks";
 export { default as QuickLinks } from "../QuickLinks";
